+++ PSDsHook/classes/DiscordEmbed.ps1
@@ -1,23 +1,33 @@
 class DiscordEmbed {
-
     [string]$title                        = [string]::Empty
     [string]$description                  = [string]::Empty
     [System.Collections.ArrayList]$fields = @()
     [string]$color                        = [DiscordColor]::New().ToString()   
     $thumbnail                            = [string]::Empty
+    $image                                = [string]::Empty
+    $author                               = [string]::Empty
+    $footer                               = [string]::Empty
+    $url                                  = [string]::Empty
 
     DiscordEmbed()
     {
         Write-Error "Please provide a title and description (and optionally, a color)!"
     }
 
-    DiscordEmbed([string]$embedTitle, [string]$embedDescription)
+    DiscordEmbed(
+        [string]$embedTitle, 
+        [string]$embedDescription
+    )
     {
         $this.title       = $embedTitle
         $this.description = $embedDescription
     }
 
-    DiscordEmbed([string]$embedTitle, [string]$embedDescription, [DiscordColor]$embedColor)
+    DiscordEmbed(
+        [string]      $embedTitle, 
+        [string]      $embedDescription, 
+        [DiscordColor]$embedColor
+    )
     {
         $this.title       = $embedTitle
         $this.description = $embedDescription
@@ -45,7 +55,55 @@ class DiscordEmbed {
         } 
         else 
         {
-            Write-Error "Did not receive a [DiscordField] object!"
+            Write-Error "Did not receive a [DiscordThumbnail] object!"
+        }
+    }
+
+    [void]AddImage($image)
+    {
+        if ($image.PsObject.TypeNames[0] -eq 'DiscordImage')
+        {
+            $this.image = $image
+        } 
+        else 
+        {
+            Write-Error "Did not receive a [DiscordImage] object!"
+        }
+    }
+
+    [void]AddAuthor($author)
+    {
+        if ($author.PsObject.TypeNames[0] -eq 'DiscordAuthor')
+        {
+            $this.author = $author
+        } 
+        else 
+        {
+            Write-Error "Did not receive a [DiscordAuthor] object!"
+        }
+    }
+
+    [void]AddFooter($footer)
+    {
+        if ($footer.PsObject.TypeNames[0] -eq 'DiscordFooter')
+        {
+            $this.footer = $footer
+        } 
+        else 
+        {
+            Write-Error "Did not receive a [DiscordFooter] object!"
+        }
+    }
+
+    [void]WithUrl($url)
+    {
+        if (![string]::IsNullOrEmpty($url))
+        {
+            $this.url = $url
+        } 
+        else 
+        {
+            Write-Error "Please provide a url!"
         }
     }