+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.4] Unreleased
+## [0.1.5] Unreleased
 
 ### Changes
 
 - Coming soon
+
+## [0.1.4] 2018-10-24
+
+### Added
+- [Author](https://discordapp.com/developers/docs/resources/channel#embed-object-embed-author-structure) object added, and can be added to the EmbedBuilder.
+- [Footer](https://discordapp.com/developers/docs/resources/channel#embed-object-embed-footer-structure) object added, and can be added to the EmbedBuilder.
+- [Image](https://discordapp.com/developers/docs/resources/channel#embed-object-embed-image-structure) object added, and can be added to the EmbedBuilder.
+- Added example using all embed objects to [examples](https://github.com/gngrninja/PSDsHook/tree/master/examples) folder.
 
 ## [0.1.3] 2018-10-20